% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplot.R
\name{oplot}
\alias{oplot}
\title{Observable Plot 'htmlwidget'}
\usage{
oplot(x = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{height, width}{height and width of sunburst 'htmlwidget' containing 'div'
specified in any valid \code{CSS} size unit.}

\item{elementId}{string id as a valid \code{CSS} element id.}

\item{spec}{\code{list} with Observable plot 'spec' in 'R' form}
}
\value{
\code{htmlwidget}
}
\description{
Simple barebones functions to create \href{https://observablehq.com/plot/}{Observable plot}
  \code{htmlwidget} from 'R' for use in 'HTML', 'rmarkdown', 'quarto', and 'Shiny'.
}
