% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplot.R
\name{oplot_dependencies}
\alias{oplot_dependencies}
\title{'oplot' 'JavaScript' Dependencies}
\usage{
oplot_dependencies()
}
\value{
\code{list} of \code{htmltools} dependencies
}
\description{
'oplot' 'JavaScript' Dependencies
}
\seealso{
\link[htmltools]{htmlDependencies}
}
